+++ src/types/SortType.ts
@@ -1,0 +1,5 @@
+export enum SortType {
+  ALPHABET,
+  LENGTH,
+  Default,
+}
